refactor: normalize index extensions, update imports
