+++ libs/shared/components/src/index.ts
@@ -1,3 +1,6 @@
 export * from './Cards';
+export * from './LinkIcon';
+export * from './Loader';
+export * from './MiddleTruncated';
+export * from './Mix';
 export * from './top-nav';
-export * from './LinkIcon/LinkIcon';
