+++ libs/shared/components/src/top-nav/index.ts
@@ -1,0 +1,2 @@
+export * from './ConnectedButton';
+export * from './Icon';
