--- libs/shared/components/src/top-nav/index.tsx
@@ -1,2 +1,0 @@
-export * from './TopNav';
-export * from './Icon';
